Added utilisation of securityManager

+++ src/M1/impl/Main.java
@@ -99,35 +99,54 @@ public class Main {
 				break;
 			}
 			case 3 : {
-				System.out.println("Veuillez rentrer le pseudo du nouveau client : ");
-				String newPseudo = in.nextLine();
-				System.out.println("Veuillez rentrer le mot de passe du nouveau client : ");
-				String newMDP = in.nextLine();
-				System.out.println("Veuillez rentrer la value que vous voulez pour ce nouveau client : ");
-				int newValue = 0;
-				while(true) {
-					try{
-						newValue = in.nextInt();
-						break;
+				if(connectionManager.checkRights(securityManager, pseudo, "")) {
+					System.out.println("Vous avez les droits pour faire ça !\n");
+					System.out.println("Veuillez rentrer le pseudo du nouveau client : ");
+					String newPseudo = in.nextLine();
+					System.out.println("Veuillez rentrer le mot de passe du nouveau client : ");
+					String newMDP = in.nextLine();
+					System.out.println("Veuillez rentrer la value que vous voulez pour ce nouveau client : ");
+					int newValue = 0;
+					while(true) {
+						try{
+							newValue = in.nextInt();
+							break;
+						}
+						catch(Exception e) {
+							System.out.println("La value doit forcément être un chiffre ou un nombre");
+						}
 					}
-					catch(Exception e) {
-						System.out.println("La value doit forcément être un chiffre ou un nombre");
-					}
+					ClientImpl c = new ClientImpl(newPseudo, newMDP, newValue);
+					database.add(newPseudo, newMDP, newValue);
+					createAndAddLinks(c);
 				}
-				ClientImpl c = new ClientImpl(newPseudo, newMDP, newValue);
-				database.add(newPseudo, newMDP, newValue);
-				createAndAddLinks(c);
+				else {
+					System.out.println("Vous n'avez pas les droits pour faire ça !!");
+				}
 				break;
+
 			}
 			case 4 : {
-				System.out.println("Veuillez entrer le pseudo du client que vous souhaiter supprimer de la database : ");
-				String p = in.nextLine();
-				database.delete(p);
+				if(connectionManager.checkRights(securityManager, pseudo, "")) {
+					System.out.println("Vous avez les droits pour faire ça !\n");
+					System.out.println("Veuillez entrer le pseudo du client que vous souhaiter supprimer de la database : ");
+					String p = in.nextLine();
+					database.delete(p);
+				}
+				else {
+					System.out.println("Vous n'avez pas les droits pour faire ça !!");
+				}
 				break;
 			}
 			case 5 : {
-				String dbAll = database.read();
-				System.out.println("Voici le contenu de la DB : \n" + dbAll + "\n");
+				if(connectionManager.checkRights(securityManager, pseudo, "")) {
+					System.out.println("Vous avez les droits pour faire ça !\n");
+					String dbAll = database.read();
+					System.out.println("Voici le contenu de la DB : \n" + dbAll + "\n");
+				}
+				else {
+					System.out.println("Vous n'avez pas les droits pour faire ça !!");
+				}
 				break;
 			}
 			case 6 : {
@@ -150,39 +169,39 @@ public class Main {
 		configClientServer.portbindingtoconfigclientserver.add(portbindingtoconfigclientserver);
 		BindingClientToConfigClientServerImpl bindingclienttoconfigclientserver = new BindingClientToConfigClientServerImpl(portbindingtoconfigclientserver, c.portclienttobinding);
 		configClientServer.addBindingclienttoconfigclientserver(bindingclienttoconfigclientserver);
-		
+
 		PortConfigClientServerToBindingImpl portconfigclientservertobinding = new PortConfigClientServerToBindingImpl();
 		configClientServer.portconfigclientservertobinding.add(portconfigclientservertobinding);
 		BindingConfigClientServerToClientImpl bindingconfigclientservertoclient = new BindingConfigClientServerToClientImpl(c.portbindingtoclient, portconfigclientservertobinding);
 		configClientServer.addBindingconfigclientservertoclient(bindingconfigclientservertoclient);
-				
+
 		PortClienttoRPCImpl portclienttorpc = new PortClienttoRPCImpl();
 		RoleRPCfromClientImpl rolerpcfromclient = new RoleRPCfromClientImpl();
 		AttachmentClienttoRPCImpl attachmentclienttorpc = new AttachmentClienttoRPCImpl(portclienttorpc, rolerpcfromclient);
 		configClientServer.attachmentclienttorpc.add(attachmentclienttorpc);				
-		
+
 		PortRPCtoServerImpl portrpctoserver = new PortRPCtoServerImpl();
 		RoleRPCtoServerImpl rolerpctoserver = new RoleRPCtoServerImpl();
 		AttachmentRPCtoServerImpl attachmentrpctoserver = new AttachmentRPCtoServerImpl(portrpctoserver, rolerpctoserver);
 		configClientServer.attachmentrpctoserver.add(attachmentrpctoserver);	
-		
+
 		RPCConnectorClientServerImpl connectorClientServer = new RPCConnectorClientServerImpl(rolerpcfromclient, rolerpctoserver);
 		configClientServer.connectors.add(connectorClientServer);
-				
+
 		PortServerToRPCImpl portservertorpc = new PortServerToRPCImpl();
 		RoleRPCfromServerImpl rolerpcfromserver = new RoleRPCfromServerImpl();
 		AttachmentServerToRPCImpl attachmentservertorpc = new AttachmentServerToRPCImpl(portservertorpc, rolerpcfromserver);
 		configClientServer.attachmentservertorpc.add(attachmentservertorpc);
-		
+
 		PortRPCtoClientImpl portrpctoclient = new PortRPCtoClientImpl();
 		RoleRPCtoClientImpl rolerpctoclient = new RoleRPCtoClientImpl();
 		AttachmentRPCToClientImpl attachmentrpctoclient = new AttachmentRPCToClientImpl(portrpctoclient, rolerpctoclient);
 		configClientServer.attachmentrpctoclient.add(attachmentrpctoclient);
-		
+
 		RPCConnectorServerClientImpl connectorServerClient = new RPCConnectorServerClientImpl(rolerpcfromclient, rolerpctoserver);
 		configClientServer.connectors.add(connectorServerClient);		
 	}
-	
+
 	public static void LancementServeur() {
 
 		int port = 2380;		
@@ -191,96 +210,96 @@ public class Main {
 			sSocket = new ServerSocket(port);
 			System.out.println("Création du serveur avec le port numéro " + port);
 			//s=sSocket.accept();
-			
-			
+
+
 			PortBindingToServerImpl portbindingtoserver = new PortBindingToServerImpl();
 			PortCMtoBindingImpl portcmtobinding = new PortCMtoBindingImpl();
 			Binding bindingcmtoserver = new BindingCMtoServerImpl(portcmtobinding, portbindingtoserver);
-			
+
 			PortBindingToCMImpl portbindingtocm = new PortBindingToCMImpl();
 			PortServerToBindingImpl portservertobinding = new PortServerToBindingImpl();
 			Binding bindingservertocm = new BindingServerToCMImpl(portbindingtocm, portservertobinding);
-			
+
 			ArrayList<PortComposantFourni> portservertorpc = new ArrayList<PortComposantFourni>();
-			
+
 			ArrayList<PortComposantRequis> portrpctoserver = new ArrayList<PortComposantRequis>();
-			
+
 			PortDBtoSQLImpl portdbtosql = new PortDBtoSQLImpl();		
 			RoleSQLfromDBImpl rolesqlfromdb = new RoleSQLfromDBImpl();
 			Attachment attachmentdbtosql = new AttachmentDBtoSQLImpl(portdbtosql, rolesqlfromdb);
-			
+
 			PortSQLtoCMImpl portsqltocm = new PortSQLtoCMImpl();
 			RoleSQLtoCMImpl rolesqltocm = new RoleSQLtoCMImpl();
 			Attachment attachmentsqltocm = new AttachmentSQLtoCMImpl(portsqltocm, rolesqltocm);
-			
+
 			SQLQueryConnectorOutImpl SQLQueryConnectorOut = new SQLQueryConnectorOutImpl(rolesqlfromdb, rolesqltocm);
 			server.connectors.add(SQLQueryConnectorOut);
-			
+
 			PortCMtoSQLImpl portcmtosql = new PortCMtoSQLImpl(); 
 			RoleSQLfromCMImpl rolesqlfromcm = new RoleSQLfromCMImpl();
 			Attachment attachmentcmtosql = new AttachmentCMtoSQLImpl(portcmtosql, rolesqlfromcm);
-						
+
 			PortSQLtoDBImpl portsqltodb = new PortSQLtoDBImpl();
 			RoleSQLtoDBImpl rolesqltodb = new RoleSQLtoDBImpl();
 			Attachment attachmentsqltodb = new AttachmentSQLtoDBImpl(portsqltodb,rolesqltodb);	
-			
+
 			SQLQueryConnectorInImpl SQLQueryConnectorIn = new SQLQueryConnectorInImpl(rolesqltodb, rolesqlfromcm);
 			server.connectors.add(SQLQueryConnectorIn);
-						
+
 			PortSQtoDBImpl portsqtodb = new PortSQtoDBImpl();
 			RoleSQtoDBImpl rolesqtodb = new RoleSQtoDBImpl();
 			Attachment attachmentsqtodb = new AttachmentSQtoDBImpl(portsqtodb, rolesqtodb);
-			
+
 			PortSMtoSQImpl portsmtosq = new PortSMtoSQImpl();
 			RoleSQfromSMImpl rolesqfromsm = new RoleSQfromSMImpl();
 			Attachment attachmentsmtosq = new AttachmentSMtoSQImpl(portsmtosq, rolesqfromsm);
-			
+
 			SecurityQueryConnectorInImpl SecurityQueryConnectorIn = new SecurityQueryConnectorInImpl(rolesqtodb, rolesqfromsm);
 			server.connectors.add(SecurityQueryConnectorIn);
-			
+
 			PortSQtoSMImpl portsqtosm = new PortSQtoSMImpl();
 			RoleSQtoSMImpl rolesqtosm = new RoleSQtoSMImpl();
 			Attachment attachmentsqtosm = new AttachmentSQtoSMImpl(portsqtosm, rolesqtosm);
-						
+
 			PortDBtoSQImpl portdbtosq = new PortDBtoSQImpl();			
 			RoleSQfromDBImpl rolesqfromdb = new RoleSQfromDBImpl();
 			Attachment attachmentdbtosq = new AttachmentDBtoSQImpl(portdbtosq, rolesqfromdb);	
-			
+
 			SecurityQueryConnectorOutImpl SecurityQueryConnectorOut = new SecurityQueryConnectorOutImpl(rolesqfromdb, rolesqtosm);
 			server.connectors.add(SecurityQueryConnectorOut);
-			
+
 			PortCRtoSMImpl portcrtosm = new PortCRtoSMImpl();
 			RoleCRtoSMImpl rolecrtosm = new RoleCRtoSMImpl();
 			Attachment attachmentcrtosm = new AttachmentCRtoSMImpl(portcrtosm, rolecrtosm);
-			
+
 			PortCMtoCRImpl portcmtocr = new PortCMtoCRImpl();
 			RoleCRfromCMImpl rolecrfromcm = new RoleCRfromCMImpl();
 			Attachment attachmentcmtocr = new AttachmentCMtoCRImpl(portcmtocr, rolecrfromcm);
-			
+
 			ClearanceRequestConnectorInImpl ClearanceRequestConnectorIn = new ClearanceRequestConnectorInImpl(rolecrfromcm, rolecrtosm);
 			server.connectors.add(ClearanceRequestConnectorIn);
-			
+
 			PortSMtoCRImpl portsmtocr = new PortSMtoCRImpl();
 			RoleCRfromSMImpl rolecrfromsm = new RoleCRfromSMImpl();
 			Attachment attachmentsmtocr = new AttachmentSMtoCRImpl(portsmtocr, rolecrfromsm);
-						
+
 			PortCRtoCMImpl portcrtocm = new PortCRtoCMImpl();
 			RoleCRtoCMImpl rolecrtocm = new RoleCRtoCMImpl();
 			Attachment attachmentcrtocm = new AttachmentCRtoCMImpl(portcrtocm, rolecrtocm);
-			
+
 			ClearanceRequestConnectorOutImpl ClearanceRequestConnectorOut = new ClearanceRequestConnectorOutImpl(rolecrtocm, rolecrfromsm);
 			server.connectors.add(ClearanceRequestConnectorOut);
-			
+
 			server = new ServerConfigurationImpl(bindingcmtoserver, bindingservertocm, 
 					portservertorpc, portrpctoserver, portservertobinding, portbindingtoserver, 
 					attachmentsqltocm, attachmentcmtosql, attachmentsqltodb, attachmentdbtosql, attachmentdbtosq, 
 					attachmentsqtodb, attachmentsqtosm, attachmentsmtosq, attachmentcrtocm, attachmentcmtocr, 
 					attachmentsmtocr, attachmentcrtosm);
-				
+
 			database = new DatabaseImpl(portdbtosql, portsqltodb, portdbtosq, portsqtodb);			
-			
+
 			connectionManager = new ConnectionManagerImpl(portsqltocm, portcmtosql, portbindingtocm, portcmtobinding, portcrtocm, portcmtocr);
-						
+
 			securityManager = new SecurityManagerImpl(portsmtosq, portsqtosm, portcrtosm, portsmtocr);
 
 		} catch (IOException e) {
